+++ src/main/java/com/fpt/sep490/service/TransactionService.java
@@ -9,7 +9,7 @@ import java.util.Set;
 public interface TransactionService {
     Transaction updateTransaction(TransactionDto transactionDto);
 
-    Set<TransactionDto> getTransactionByReceiptId(int receiptId);
+    Set<TransactionDto> getTransactionByReceiptId(long receiptId);
 
     List<Transaction> getAllTransaction();
